Enable connection removal and bug fixes
BUG - where null is not shown on text area, because the initial data was already null, no update is made to the text area. Call calculate when connection is added and removed, because data can be null and won't trigger a new calculation
BUG - where text is not removed from text field, after removing a connection. Resetting text area when connection is removed.
REFACTOR - moved workspace reset from actions as public method into the workspace class